C# Unt32 -> UInt64, int -> ulong

+++ inc/inc.cs
@@ -3,7 +3,7 @@ public class Program
 {
     public static void Main(string[] args)
     {
-        for (int i = 1; i <= Int32.Parse(args[0]); i++)
+        for (ulong i = 1; i <= UInt64.Parse(args[0]); i++)
         {
 //            Console.WriteLine(i);
         }
